Fixed comparison warnings in audio_stream and video_stream

+++ torchvision/csrc/io/decoder/audio_stream.cpp
@@ -7,12 +7,14 @@ namespace ffmpeg {
 
 namespace {
 bool operator==(const AudioFormat& x, const AVFrame& y) {
-  return x.samples == y.sample_rate && x.channels == y.channels &&
+  return x.samples == static_cast<size_t>(y.sample_rate) &&
+      x.channels == static_cast<size_t>(y.channels) &&
       x.format == y.format;
 }
 
 bool operator==(const AudioFormat& x, const AVCodecContext& y) {
-  return x.samples == y.sample_rate && x.channels == y.channels &&
+  return x.samples == static_cast<size_t>(y.sample_rate) &&
+      x.channels == static_cast<size_t>(y.channels) &&
       x.format == y.sample_fmt;
 }
